ref(preprod): Move installableQueryUtils to views/preprod/utils
Better home alongside the other preprod utility modules. No behavior
change.
