+++ static/app/views/releases/list/mobileBuilds.tsx
@@ -7,7 +7,6 @@ import {Stack} from '@sentry/scraps/layout';
 import {LoadingError} from 'sentry/components/loadingError';
 import {LoadingIndicator} from 'sentry/components/loadingIndicator';
 import {normalizeDateTimeParams} from 'sentry/components/pageFilters/parse';
-import {getUpdatedQueryForDisplay} from 'sentry/components/preprod/installableQueryUtils';
 import {
   getPreprodBuildsDisplay,
   PreprodBuildsDisplay,
@@ -23,6 +22,7 @@ import {useLocation} from 'sentry/utils/useLocation';
 import {useNavigate} from 'sentry/utils/useNavigate';
 import {usePreprodBuildsAnalytics} from 'sentry/views/preprod/hooks/usePreprodBuildsAnalytics';
 import {buildDetailsApiOptions} from 'sentry/views/preprod/utils/buildDetailsApiOptions';
+import {getUpdatedQueryForDisplay} from 'sentry/views/preprod/utils/installableQueryUtils';
 
 import {MobileBuildsChart} from './mobileBuildsChart';
